backend/requirements.txt
Django>=5.0
psycopg[binary]
djangorestframework
django-environ
django-cors-headers
requests
pytest
pytest-django
pytest-mock
flake8
black
